+++ Database/DatabaseSetup.cs
@@ -16,7 +16,7 @@ class DatabaseSetup
 
     private void CreateTableComputer()
     {
-        var connection = new SqliteConnection("Data Source=database.db");
+        var connection = new SqliteConnection(databaseConfig.ConnectionString);
         connection.Open();
 
         var command = connection.CreateCommand();
