+++ Repositories/ComputerRepository.cs
@@ -1,4 +1,5 @@
 using LabManager.Models;
+using LabManager.Database;
 using Microsoft.Data.Sqlite;
 
 namespace LabManager.Repositories;
@@ -6,11 +7,15 @@ namespace LabManager.Repositories;
 class ComputerRepository
 {
     private DatabaseConfig databaseConfig;
+
+
+    public ComputerRepository(DatabaseConfig databaseConfig) => this.databaseConfig = databaseConfig;
+
     public List<Computer> GetAll()
     {
         var computers = new List<Computer>();
         
-        var connection = new SqliteConnection("Data Source=database.db");
+        var connection = new SqliteConnection(databaseConfig.ConnectionString);
         connection.Open();
 
         var command = connection.CreateCommand();
@@ -29,5 +34,21 @@ class ComputerRepository
         connection.Close(); 
 
         return computers;
+    }
+
+
+    public void Save (Computer computer)    //se passa o obj como parâmetro, pois eram muitos atributos
+    {
+        var connection = new SqliteConnection(databaseConfig.ConnectionString);
+            connection.Open();
+
+            var command = connection.CreateCommand();
+            command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor);";
+            command.Parameters.AddWithValue("$id", computer.Id);
+            command.Parameters.AddWithValue("$ram", computer.Ram);
+            command.Parameters.AddWithValue("$processor", computer.Processor);
+
+            command.ExecuteNonQuery();
+            connection.Close();  
     }
 }
